# Tune quant dispatch