Tune quant dispatch conditions to qmv vs qmm.

The optimal is both machine and matrix size dependent unfortunately so this is only an approximation. It's conservative in that we continue to route to the qmm unless it's pretty clear that there is a speedup to be had.

In the plots below:
- x-axis is batch size
- y-axis is time in milliseconds
- dashed line is qmm
- solid line is qmv

Here's a benchmark comparing the two at size = 4096 on a few machines:

<img width="1037" alt="Screenshot 2025-04-01 at 8 08 00 AM" src="https://github.com/user-attachments/assets/7b8327ea-4a31-4a46-94fa-2763eb42b356" />

Another benchmark for just M2 Ultra for 3 different sized matrices:

<img width="1235" alt="Screenshot 2025-04-01 at 8 09 12 AM" src="https://github.com/user-attachments/assets/9c1e1398-e819-4583-b290-c5f124994e4c" />

Same benchmark on M3 Max:

<img width="1227" alt="Screenshot 2025-04-01 at 8 09 35 AM" src="https://github.com/user-attachments/assets/90067895-403f-45f1-b342-47ffa6ed66ec" />