@awni 2025-04-01
For Mistral 7B batched we get a nice speedup especially at small batch sizes:

Batch size | Pre ms | Post ms
----- | ---- | -----
6 | 31.971 | 19.400
8 | 31.817 | 25.431
10 |31.416| 31.274


Benchmark:

```python
import time
import mlx.core as mx
from mlx_lm import load

model, _ = load("mlx-community/mistral-7B-Instruct-v0.3-4bit")

for batch_size in [6, 8, 10, 12, 14, 20]:
    prompt = mx.array([[0] * batch_size])

    for _ in range(10):
        mx.async_eval(model(prompt))

    tic = time.time()
    for _ in range(10):
        mx.async_eval(model(prompt))
    toc = time.time()
    ms = 1000 * (toc - tic) / 10
    print(f"{ms:.3f}")
```