@awni 2025-04-01
It's true I didn't tune for those at all. I just checked `qmv` and on m4 max the crossing point is 14 (for size == 4352), so close to `qmv_fast`. I'm gues quad will be quite different. If anything I would expect the crossing point to be further out since the matrices are probably quite small with that one.